Fix upgrade to 1.16 forcing recreation.
When upgrading a cluster from 1.15 with legacy_security_groups = true
the description change to the legacy control_plane explaining why
this still existed forces the security group to be recreated.
This forces the cluster and a number of other resources to also be
recreated.

This change reverts the description to it's previous value to avoid that
recreation.