[IAM module] Fix Invalid depends_on reference error
The following error occurs while terraform plan is executed.

```
terraform plan
Acquiring state lock. This may take a few moments...
Releasing state lock. This may take a few moments...

Error: Invalid depends_on reference

  on .terraform/modules/eks_iam/modules/iam/outputs.tf line 7, in output "config":
   7:     aws_iam_role_policy_attachment.eks_worker_node.policy_arn,

References in depends_on must be to a whole object (resource, etc), not to an
attribute of an object.

Error: Invalid depends_on reference

  on .terraform/modules/eks_iam/modules/iam/outputs.tf line 8, in output "config":
   8:     aws_iam_role_policy_attachment.eks_cni.policy_arn,

References in depends_on must be to a whole object (resource, etc), not to an
attribute of an object.

Error: Invalid depends_on reference

  on .terraform/modules/eks_iam/modules/iam/outputs.tf line 9, in output "config":
   9:     aws_iam_role_policy_attachment.ecr.policy_arn,

References in depends_on must be to a whole object (resource, etc), not to an
attribute of an object.

Error: Invalid depends_on reference

  on .terraform/modules/eks_iam/modules/iam/outputs.tf line 10, in output "config":
  10:     aws_iam_role_policy_attachment.eks_service_policy.policy_arn,

References in depends_on must be to a whole object (resource, etc), not to an
attribute of an object.

Error: Invalid depends_on reference

  on .terraform/modules/eks_iam/modules/iam/outputs.tf line 11, in output "config":
  11:     aws_iam_role_policy_attachment.eks_cluster_policy.policy_arn,

References in depends_on must be to a whole object (resource, etc), not to an
```

To fix this invalid dependency errors, we have to pass objects instead
of resouces(policy_arn.)