use sts-regional-entpoints=true in eks pod identity webhook
When using IAM permissions at the pod level, in most cases it is desirable to use the sts regional endpoint not the global endpont.
Using regional AWS STS endpoints instead of the global endpoint to reduce latency, build in redundancy, and increase session token validity when pods execute sts:AssumeRoleWithWebIdentity for IRSA.

When using IAM permissions at the pod level, in most cases it is desirable to use the sts regional endpoint not the global endpont.