Seperate karpenter config to a submodule
It seemed a bit weird that we did a lot of work to remove all of the
non-core functions from the module, just to add the support for
karpenter as a core part of the module.

Moving support for Karpenter to a submodule doesn't make configuring
a cluster much more dificult, but does reduce the complexity of the
core module, and hopefully should make ongoing maintince a little
simpler.  Especially if we decided to use some other node provisioning
tool in the future.