Bump aws provider version to 5
This change bumps aws provider version to 5 for the cluster sub module
and removes version constraints of "< 5.0.0" for other sub mosules to
allow the cluster sub module to actually use aws provider newer than 5.

This change should help people to upgrade to version 1.25. Currently, in
an environment that manages multiple EKS cluster with a single terraform
state, it is hard to upgrade to version 1.25 because version 1.25 of
this terraform module implicitly requires aws provider newer than 5.