Port asg_node_group to v1.25
* We have some legacy clusters on 1.24 that need to be upgraded before
extended support ends.
* They are still using cluster-autoscaler (not Karpenter)
* The simplest upgrade procedure is to forward-port the asg_node_group
  module.

Note: this module adds a node_instance_profile variable, as cluster
config no longer exposes an instance profile for the node group.

To avoid any issues this needs to be set to the same value as you
were using in 1.24 iam_config

Also note that the IAM role of nodes in the node group will need to be
manually added to the clusters `aws_auth_role_map` as this is no longer
defaulted.